Add login with FormsAuthentication
Convert pcp login page to general login page for all users.
Edit Web.config to ensure authentication of all pages is used.
Move stylesheets to a directory so it can be accessed without
authentication (so that the login page is styled).
Add username and log out button to master page, so that all pages show
it.